feat: remove 'createintegrated object' page and related code (#763)
