--- app/views/AddNewComponentAtlasView/common/constants.ts
@@ -1,3 +1,0 @@
-export const FIELD_NAME = {
-  TITLE: "title",
-} as const;
