--- app/views/AddNewComponentAtlasView/common/entities.ts
@@ -1,4 +1,0 @@
-import { InferType } from "yup";
-import { newComponentAtlasSchema } from "./schema";
-
-export type NewComponentAtlasData = InferType<typeof newComponentAtlasSchema>;
