+++ app/views/ComponentAtlasView/common/constants.ts
@@ -1,6 +1,4 @@
-import { FIELD_NAME as NEW_COMPONENT_ATLAS_FIELD_NAME } from "../../AddNewComponentAtlasView/common/constants";
-
 export const FIELD_NAME = {
-  ...NEW_COMPONENT_ATLAS_FIELD_NAME,
   SOURCE_DATASET_IDS: "sourceDatasetIds",
+  TITLE: "title",
 } as const;
